+++ src/pages/ErrorPage/ErrorPage.scss
@@ -1,0 +1,34 @@
+@import '../../_utils';
+
+.errorMobile {
+    .errorTop {
+    height: 20vh;
+    }
+}
+
+.errorTop {
+    width: 100%;
+    height: 40vh;
+    margin-bottom: rem(20px);
+    display: flex;
+    overflow: hidden;
+    position: relative;
+    background-image: url($error-bg);
+    background-color: $green-light;
+    background-size: cover;
+    background-position: center;
+}
+
+.errorWrapper {
+    h2 {
+        font-family: $mono;
+        text-align: center;
+        margin-top: rem(40px);
+        margin-bottom: rem(20px);
+        padding: rem(10px);
+    }
+
+    a {
+        text-decoration: underline;
+    }
+}
